Refactored Cilium Operations
